Add files via upload
Here are three tasks that are done by using arithmetic operations, In KilotoGrams this program will take no of kilos from user and convert kilos into grams.
In SecondsToMints the user has to give seconds as input and program will convert seconds into mints. In the marksheet I have explained a simple and basic level marksheet which is taking 5 subjects marks as input and showing total, obtained and percentage.